Styled main page and moved components into mainComponents folder
