+++ client/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import {PricingComponent} from './pricing/pricing.component';
 import {DownloadComponent} from './download/download.component';
-import {MainComponent} from './main/main.component';
+import {MainComponent} from './mainComponents/main/main.component';
 import {SubjectsComponent} from './subjects/subjects.component';
