+++ client/src/app/app.module.ts
@@ -3,12 +3,12 @@ import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header.component';
+import { HeaderComponent } from './mainComponents/header/header.component';
 import { PricingComponent } from './pricing/pricing.component';
 import { DownloadComponent } from './download/download.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FooterComponent } from './footer/footer.component';
-import { MainComponent } from './main/main.component';
+import { FooterComponent } from './mainComponents/footer/footer.component';
+import { MainComponent } from './mainComponents/main/main.component';
 import { SubjectsComponent } from './subjects/subjects.component';
 import { HttpClientModule } from '@angular/common/http';
