+++ client/src/app/mainComponents/footer/footer.component.scss
