+++ client/src/app/mainComponents/footer/footer.component.spec.ts
