+++ client/src/app/mainComponents/footer/footer.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss','../../assets/css/style-starter.css']
+  styleUrls: ['./footer.component.scss','../../../assets/css/style-starter.css']
 })
 export class FooterComponent implements OnInit {
