+++ client/src/app/mainComponents/main/main.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-main',
   templateUrl: './main.component.html',
-  styleUrls: ['./main.component.scss' , '../../assets/css/style-starter.css']
+  styleUrls: ['./main.component.scss' , '../../../assets/css/style-starter.css']
 })
 export class MainComponent implements OnInit {
